+++ tests/cmd/stress.go
@@ -4,6 +4,8 @@ import (
 	"context"
 	"os"
 
+	broker "gitlab.com/ConsenSys/client/fr/core-stack/orchestrate.git/v2/pkg/broker/sarama"
+
 	traefiklog "github.com/containous/traefik/v2/pkg/log"
 	"github.com/spf13/cobra"
 	"gitlab.com/ConsenSys/client/fr/core-stack/orchestrate.git/v2/pkg/utils"
@@ -20,6 +22,7 @@ func NewRunStressTestCommand() *cobra.Command {
 
 	// Register Stress flag
 	stress.InitFlags(runCmd.Flags())
+	broker.ConsumerGroupName(runCmd.Flags())
 
 	return runCmd
 }