+++ tests/service/e2e/exported.go
@@ -110,7 +110,7 @@ func initComponents(ctx context.Context) {
 		},
 		// Initialize ConsumerGroup
 		func() {
-			broker.InitConsumerGroup(ctx, "group-cucumber")
+			broker.InitConsumerGroup(ctx, viper.GetString(broker.ConsumerGroupNameViperKey))
 		},
 		// Initialize Handlers
 		func() {